Changed API signature for set/get seed functions

+++ rali/include/parameter_factory.h
@@ -53,8 +53,8 @@ public:
     static ParameterFactory* instance();
     ~ParameterFactory();
     void renew_parameters();
-    void set_seed(long long unsigned seed);
-    unsigned long long get_seed();
+    void set_seed(unsigned seed);
+    unsigned get_seed();
 
     template<typename T>
     Parameter<T>* create_uniform_rand_param(T start, T end){
